chore: refactor two factor auth screen [#181358246]
- use numberic input for code
- reduce code
- remove snapshot
