+++ web/src/screens/ProfileTwoFactorAuthScreen/ProfileTwoFactorAuthScreen.postcss
@@ -1,0 +1,7 @@
+.twoFaSteps {
+  padding-left: calc(var(--gap) * 3);
+}
+
+.twoFaQr {
+  width: 120px;
+}
