+++ web/src/screens/style.pcss
@@ -4,7 +4,6 @@
 @import './History/HistoryTab.pcss';
 @import './ConfirmScreen/ConfirmScreen.pcss';
 @import './ProfileScreen/ProfileScreen.pcss';
-@import './ProfileTwoFactorAuthScreen/ProfileTwoFactorAuthScreen.pcss';
 @import './OrdersTabScreen/OrdersTabScreen.pcss';
 @import './RestrictedScreen/RestrictedScreen.pcss';
 @import './MaintenanceScreen/MaintenanceScreen.pcss';
