--- web/src/screens/tests/ProfileTwoFactorAuthScreen.test.tsx
@@ -1,17 +1,0 @@
-import { shallow } from 'enzyme';
-import { TestComponentWrapper } from 'src/lib/test';
-import { ProfileTwoFactorAuthScreen } from '..';
-
-const setup = () =>
-  shallow(
-    <TestComponentWrapper>
-      <ProfileTwoFactorAuthScreen />
-    </TestComponentWrapper>,
-  );
-
-describe('ProfileTwoFactorAuthScreen test', () => {
-  it('should render', () => {
-    const wrapper = setup().render();
-    expect(wrapper).toMatchSnapshot();
-  });
-});
